Move date utils functions to handlers file
