+++ app/utils/handlers.js
@@ -29,3 +29,8 @@ export function openLink(e: SyntheticEvent<HTMLAnchorElement>) {
   // $FlowFixMe
   shell.openExternal(e.target.href);
 }
+
+export const toISODateString = (date: Date) =>
+  `${date.toISOString().split('.')[0]}Z`;
+
+export const formatDate = (ISODate: string) => ISODate.split('T')[0];
