--- app/utils/index.js
@@ -1,4 +1,0 @@
-export const toISODateString = (date: Date) =>
-  `${date.toISOString().split('.')[0]}Z`;
-
-export const formatDate = (ISODate: string) => ISODate.split('T')[0];
